requirements.txt
aiogram==3.0.0b3
python-dotenv==0.20.0